refactor: extract OidCache from Connection
Move enum OID caching into a dedicated OidCache type. Its resolve()
method takes &Client so Connection doesn't need to lend both &mut self
fields simultaneously.
